git-try-push: skip initial push without force when `no_lease`
The pull-and-rebase step that this action does upon failure of the
initial `git push` (without force) destroys the changes we were trying
to force push in the first place.

Let's fix that by pushing with force right from the beginning. We've
already decided we're willing to destroy changes on the target branch,
so let's skip the initial `git push` which we already know is going to
fail.

See Homebrew/homebrew-core#126950 for additional context.