check-commit-format: return failing status for non-compliant commits
Let's make this check return a failing status whenever the PR has
commits that don't follow our style guide. This will be a stronger
nudge to contributors to fix up their PR branch.

This currently isn't a required check in Homebrew/core, so it shouldn't
affect the rest of the CI run or our ability to merge PRs.