check-commit-format: add autosquash label when needed
This adds the `autosquash` label to PRs that used to automatically be
autosquashed.

Also, while we're here, let's set a failing commit status for:
1. PRs with merge commits
2. PRs with commits that modify multiple formulae

Autosquash never handled PRs with these before, and maintainers end up
either having to ask users to fix their branch or fixing it up
themselves. Let's try to avoid that by nudging the user to fix it.